Use long bit offsets when reading/writing fixed bit width files
RB=533339
BUG=
G=pinot-dev-reviewers
R=kgopalak,xiafu,jfim,dpatel,mshrivas
A=kgopalak

+++ pinot-core/src/main/java/com/linkedin/pinot/core/index/writer/impl/FixedBitWidthRowColDataFileWriter.java
@@ -138,7 +138,7 @@ public class FixedBitWidthRowColDataFileWriter implements Closeable {
    */
   public void setInt(int row, int col, int val) {
     assert val >= minValues[col]  && val <= maxValues[col];
-    int bitOffset = rowSizeInBits * row + columnOffsetsInBits[col];
+    long bitOffset = ((long) rowSizeInBits) * row + columnOffsetsInBits[col];
     val = val + offsets[col];
     for (int bitPos = colSizesInBits[col] - 1; bitPos >= 0; bitPos--) {
       if ((val & (1 << bitPos)) != 0) {
